+++ R/helpers.r
@@ -428,3 +428,21 @@ namedNames <- function(g) {
   n
 }
 
+embedKnnGraphUmap <- function(knn.graph, k=NULL, ...) {
+  if (!requireNamespace("uwot", quietly=T))
+    stop("You need to install package 'uwot' to be able to use UMAP embedding.")
+
+  adj.mat <- igraph::as_adj(knn.graph, attr="weight") %>% as("dgTMatrix")
+  vals.per.col <- split(setNames(adj.mat@x, adj.mat@i + 1), adj.mat@j + 1)
+  k.min <- sapply(vals.per.col, length) %>% min()
+  k <- if (is.null(k)) k.min else min(k, k.min)
+  
+  knns <- lapply(vals.per.col, function(x) sort(x, decreasing=T)[1:k])
+  knn.ids <- sapply(knns, function(x) as.integer(names(x))) %>% t()
+  knn.sims <- do.call(rbind, knns)
+  knn.dists <- 1 - knn.sims / max(knn.sims)
+  
+  umap <- uwot::umap(data.frame(x=rep(0, nrow(knn.ids))), nn_method=list(idx=knn.ids, dist=knn.dists), ...)
+  rownames(umap) <- colnames(adj.mat)
+  return(umap)
+}